requirements.txt
opencv-python==4.4.0.44
tflearn==0.3.2
paho-mqtt==1.5.1
Pillow==7.1.2